Fixed link suppression
Previously subsequent links in paragraphs would be suppressed.  As per
issue #377 this meant that links in lists were not being handled
correctly.  In addition, links in the master indices were being
suppressed when they really should not be.

New rule is that if references to roles, states, or properties are in a
table or a dl (definition list), then they are left as links.
Otherwise, only the first reference to a role, state, or property in a
section will be a link.  Subsequent ones will be just be terms that are
not linked.

Also cleaned up some JS warnings to make the code cleaner (no effect on
execution) and wrapped the links to roles, states, and properties in
code elements so that they are visually the same as their non-linked
counterparts.