Rely more upon Specref
The following items are in Specref, but we have slightly different names
for them locally:

* Our "ACCNAME-AAM": https://www.specref.org/?q=ACCNAME-AAM-1.1
* Our "ARIA-PRACTICES": https://www.specref.org/?q=WAI-ARIA-PRACTICES-1.1
* Our "CORE-AAM": https://www.specref.org/?q=CORE-AAM-1.1
* Our "DPUB-ARIA": https://www.specref.org/?q=DPUB-ARIA-1.0
* Our "GRAPHICS-AAM": https://www.specref.org/?q=GRAPHICS-AAM-1.0
* Our "GRAPHICS-ARIA": https://www.specref.org/?q=GRAPHICS-ARIA-1.0
* Our "HTML-AAM": https://www.specref.org/?q=HTML-AAM-1.0
* Our "SVG-AAM": https://www.specref.org/?q=SVG-AAM-1.0
* Our "SVG1": https://www.specref.org/?q=SVG

By replacing local entries with aliasOf references to Specref entries,
we don't have to change any of our specs, and can still get the benefit
of letting Specref do the maintenance work.

Ultimately, we should see which of these references are still being used
in any of our specs, update those specs to use versioned references, and
then we can remove these entries entirely.