Don't make issue text all uppercase
The `issue` class in ReSpec is applied to an entire `div` for a titled issue (which is like a note or editor's note, but labelled issue, with a red background, and included in the [Issue Summary](https://github.com/w3c/respec/wiki/issue-summary) index).

Making the entire issue text uppercase is unreadable. The labelling and color formatting should be enough to draw attention to an issue.