Fix "Platform Mapping Maintainers" transformation for AAMs
The transformation of "Authors" to "Platform Mapping Maintainers" is
done by finding the dt element in the head and matching on the string
"Authors:". Due to some change somewhere (perhaps ReSpec), the "Authors:"
string is now surrounded by whitespace. Trimming the whitespace before
doing the string comparison fixes the problem.

Addresses github issue #25